requirements.txt
typer
mysqlclient
python-dotenv
ollama